pyproject.toml
[tool.poetry]
name = "signup"
version = "0.1.0"
description = ""
authors = ["Felipe Silva"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^1.2.2"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.8"
psycopg2-binary = "^2.9.9"
aiormq = "^6.8.0" # https://pypi.org/project/aiormq/
uvicorn = "^0.30.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
